requirements.txt
asgiref==3.2.10
Django==3.1.1
django-material-admin==1.7.8
djangorestframework==3.11.1
mysqlclient==1.4.6
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
